backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
email-validator
transformers
torch
pandas
python-multipart